Some changes in the structure: GradedModulePresentationsForCAP
Now the function TRY_TO_ENHANCE_HOMALG_GRADED_RING_WITH_RANDOM_FUNCTIONS will be
called in the category constructor of graded left/right presentations over a
homalg ring R.

This function tries to add 3 functions to compute

    * random element with a given degree
    * random matrix between two graded free presentations (matrix of
      some left presentation)
    * random matrix for a left presentation

over the graded ring R. If they are added then the function
returns true, otherwise false.

If the output is true then the added functions are

    R!.random_graded_element_func (degree)

    R!.random_matrix_between_free_left_presentations_func
    (degrees_of_source, degrees_of_range)

    R!.random_matrix_for_left_presentation_func ( nr_rows, nr_cols )

The random methods will be added to the category of graded left/right presentations
only if the previous components of R are bound.

(I.e., It is possible to add them by hand before contructing the category)