Install logical implications in Finalize
This allows us to suspend/resume method reordering once instead of multiple
times per category. The only regression is that this way all categorical
properties have to be set before calling Finalize, but setting properties
afterwards is a bad idea anyway because then derivations specific to this
property cannot be triggered anymore.