Allow to distinguish between (small) integers and big integers for Julia
Ideally, we would declare `IsBigInt` as a filter different from `IsInt` to
properly distinguish the two in the type system of CompilerForCAP. However,
this would make calling `InstallMethod` with filter `IsBigInt` impossible
because we cannot force GAP's integers to lie in `IsBigInt`. Additionally,
we would have to dynamically decide if literal integers should be
interpreted as `IsInt` or as `IsBigInt` depending on the context in the
type system of CompilerForCAP. So for now we just make `IsBigInt` a synonym
of `IsInt`.