Refactor implications of `IsBicartesian(Co)ClosedCategory`
Make `IsBicartesianClosedCategory` equivalent to
`IsBicartesianCategory and IsCartesianClosedCategory`.
In particular, the implication
`IsBicartesianClosedCategory => IsBicartesianCategory`
is now explicit and does not need the implication
`IsCartesianCategory and IsCocartesianCategory => IsBicartesianCategory`.
This is useful for Julia.

Also, due to a typo, the implication
`IsBicartesianClosedCategory => IsCartesianCategory`
was installed twice. The second installation should have been
`IsBicartesianCoclosedCategory => IsCartesianCategory`.